requirements.txt
numpy
pyserial
pyqtgraph
pygatt